refactor of comments plugin to use a set of base plugin functions
