+++ src/helpers/createMaterialEditor.js
@@ -2,9 +2,10 @@ import { createEditor } from 'slate'
 import { withReact } from 'slate-react'
 import { withHistory } from 'slate-history'
 
+//Material Editor Plugins
+import withBase from '../plugins/withBase'
 import withMarks  from '../plugins/withMarks'
 import withBlocks from '../plugins/withBlocks'
-import withSyncedElements from '../plugins/withSyncedElements'
 import withComments from '../plugins/withComments'
 import withEndnotes from '../plugins/withEndnotes'
 
@@ -12,6 +13,7 @@ import withEndnotes from '../plugins/withEndnotes'
  * Creates a RichText editor.
  * 
  * Includes the following plugins
+ *  - withBase
  *  - withBlocks
  *  - withMarks
  *  - withHistory
@@ -24,9 +26,9 @@ export default function createMaterialEditor(editorId = "default") {
   const editor = 
   withEndnotes( 
     withComments(
-      withSyncedElements(
-        withBlocks( 
-          withMarks(
+      withBlocks( 
+        withMarks(
+          withBase( 
             withHistory(
               withReact(
                 createEditor()
